requirements.txt
pygame-ce
animsprite_pygame
pygame_widgets